backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["evgestrogan <evgestrogan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
uvicorn = "^0.17.5"
dnspython = "^2.2.1"
pydantic = {extras = ["email"], version = "^1.9.0"}
python-dotenv = "^0.19.2"
python-jose = "^3.3.0"
PyYAML = "^6.0"
sqlmodel = "^0.0.8"
asyncpg = "^0.26.0"

[tool.poetry.dev-dependencies]
httpx = "^0.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
